+++ package.json
@@ -24,18 +24,18 @@
     "precommit": "lint-staged && npm run lint"
   },
   "dependencies": {
-    "@nestjs/apollo": "^10.0.16",
-    "@nestjs/common": "^8.0.0",
+    "@nestjs/apollo": "^12.2.2",
+    "@nestjs/common": "^10.4.13",
     "@nestjs/config": "^2.1.0",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/graphql": "^10.0.16",
+    "@nestjs/core": "^10.4.13",
+    "@nestjs/graphql": "^12.2.2",
     "@nestjs/mongoose": "^9.1.1",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/swagger": "^5.2.1",
-    "@nestjs/typeorm": "^8.1.4",
+    "@nestjs/platform-express": "^10.4.13",
+    "@nestjs/swagger": "^8.1.0",
+    "@nestjs/typeorm": "^7.1.5",
     "apollo-server-express": "^3.9.0",
     "class-transformer": "^0.5.1",
-    "class-validator": "^0.13.2",
+    "class-validator": "^0.14.1",
     "graphql": "^16.5.0",
     "mongoose": "^6.3.8",
     "pg": "^8.7.3",
@@ -46,9 +46,9 @@
     "typeorm": "^0.3.6"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.0.0",
+    "@nestjs/cli": "^10.4.8",
     "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
+    "@nestjs/testing": "^7.5.5",
     "@types/express": "^4.17.13",
     "@types/jest": "27.5.0",
     "@types/node": "^16.0.0",
