+++ src/app.module.ts
@@ -2,14 +2,14 @@ import { Module } from '@nestjs/common';
 import { ConfigurationModule } from './configuration/configuration.module';
 import { AuthorController } from './controllers';
 import { DataServicesModule } from './services/data-services/data-services.module';
-import { AuthorUseCasesModule } from './use-cases/author/author-use-cases.module';
+import { FeatureflagUseCasesModule } from './use-cases/featureflag/featureflag-use-cases.module';
 import { GraphQLModule } from '@nestjs/graphql';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 
 @Module({
   imports: [
     ConfigurationModule,
-    AuthorUseCasesModule,
+    FeatureflagUseCasesModule,
     DataServicesModule,
     GraphQLModule.forRoot<ApolloDriverConfig>({
       driver: ApolloDriver,
