--- src/core/entities/author.entity.ts
@@ -1,11 +1,0 @@
-export class Author {
-  _id: string;
-  firstName: string;
-  lastName: string;
-}
-
-export class AuthorPg {
-  _id: string;
-  firstName: string;
-  lastName: string;
-}
