+++ src/use-cases/featureflag/featureflag-factory.service.ts
@@ -1,18 +1,19 @@
 import { Injectable } from '@nestjs/common';
-import { Author } from '../../core/entities';
+import { FeatureFlag } from '../../core/entities';
+import { FeatureFlagResponse } from '../../core/entities';
 import { CreateAuthorDto, UpdateAuthorDto } from '../../core/dtos';
 
 @Injectable()
-export class AuthorFactoryService {
-  createNewAuthor(createAuthorDto: CreateAuthorDto) {
-    const newAuthor = new Author();
+export class FeatureFlagFactoryService {
+  createNewFeatureflag(createFeatureflagDto: CreateFeatureflagDto) {
+    const newFeatureflag = new FeatureFlag();
     newAuthor.firstName = createAuthorDto.firstName;
     newAuthor.lastName = createAuthorDto.lastName;
 
     return newAuthor;
   }
 
-  updateAuthor(updateAuthorDto: UpdateAuthorDto) {
+  updateFeatureflag(updateFeatureflagDto: UpdateFeatureflagDto) {
     const newAuthor = new Author();
     newAuthor.firstName = updateAuthorDto.firstName;
     newAuthor.lastName = updateAuthorDto.lastName;
